+++ sdk/cosmos/azure_data_cosmos_driver/src/models/mod.rs
@@ -41,7 +41,10 @@ pub use account_reference::{AccountReference, AccountReferenceBuilder, Credentia
 pub use activity_id::ActivityId;
 pub use connection_string::ConnectionString;
 pub(crate) use consistency_level::DefaultConsistencyLevel;
-pub use cosmos_headers::{CosmosRequestHeaders, CosmosResponseHeaders};
+pub use cosmos_headers::{
+    AutoscaleAutoUpgradePolicy, AutoscaleThroughputPolicy, CosmosRequestHeaders,
+    CosmosResponseHeaders, OfferAutoscaleSettings,
+};
 pub use cosmos_operation::CosmosOperation;
 pub use cosmos_resource_reference::CosmosResourceReference;
 pub(crate) use cosmos_resource_reference::ResourcePaths;